Conflicts:
	app/controllers/SettingsController.java

+++ app/controllers/EnrollController.java
@@ -34,9 +34,11 @@ public class EnrollController extends BaseController {
         List<Exam> exams = Ebean.find(Exam.class)
                 .fetch("creator", "firstName, lastName")
                 .fetch("examLanguages")
+                .fetch("examOwners", "firstName, lastName")
+                .fetch("course", "code, name")
                 .where()
                 .eq("course.code", code)
-                .eq("state", "PUBLISHED")
+                .eq("state", Exam.State.PUBLISHED)
                 .ge("examActiveEndDate", new Date())
                 .findList();
 
@@ -104,8 +106,8 @@ public class EnrollController extends BaseController {
                 .isNull("reservation")
                 .endJunction()
                 .disjunction()
-                .eq("exam.state", "PUBLISHED")
-                .eq("exam.state", "STUDENT_STARTED")
+                .eq("exam.state", Exam.State.PUBLISHED)
+                .eq("exam.state", Exam.State.STUDENT_STARTED)
                 .endJunction()
                 .findList();
         if (enrolments.isEmpty()) {
@@ -116,8 +118,13 @@ public class EnrollController extends BaseController {
 
     @Restrict({@Group("ADMIN"), @Group("STUDENT")})
     public Result removeEnrolment(Long id) {
-        ExamEnrolment enrolment = Ebean.find(ExamEnrolment.class, id);
-        if (enrolment.getReservation() != null) {
+        ExamEnrolment enrolment = Ebean.find(ExamEnrolment.class).fetch("exam")
+                .where().idEq(id).findUnique();
+        // Disallow removing enrolments to private exams created automatically for student
+        if (enrolment.getExam().isPrivate()) {
+            return forbidden();
+        }
+        if (enrolment.getReservation() != null ) {
             return forbidden("sitnet_cancel_reservation_first");
         }
         enrolment.delete();
@@ -137,11 +144,12 @@ public class EnrollController extends BaseController {
         return ok();
     }
 
-    private Result doCreateEnrolment(Long eid, Long uid) {
+    private Result doCreateEnrolment(Long eid, Long uid, ExamExecutionType.Type type) {
         User user = uid == null ? getLoggedUser() : Ebean.find(User.class, uid);
         Exam exam = Ebean.find(Exam.class)
                 .where()
                 .eq("id", eid)
+                .eq("executionType.type", type.toString())
                 .findUnique();
         if (exam == null) {
             return notFound("sitnet_error_exam_not_found");
@@ -159,7 +167,7 @@ public class EnrollController extends BaseController {
                 .disjunction()
                 .conjunction()
                 .eq("exam.parent.id", exam.getId())
-                .eq("exam.state", Exam.State.STUDENT_STARTED.toString())
+                .eq("exam.state", Exam.State.STUDENT_STARTED)
                 .endJunction()
                 .endJunction()
                 .endJunction()
@@ -172,10 +180,10 @@ public class EnrollController extends BaseController {
                 return forbidden("sitnet_error_enrolment_exists");
             } else if (reservation.toInterval().contains(AppUtil.adjustDST(DateTime.now(), reservation))) {
                 // reservation in effect
-                if (exam.getState().equals(Exam.State.STUDENT_STARTED.toString())) {
+                if (exam.getState() == Exam.State.STUDENT_STARTED) {
                     // exam for reservation is ongoing
                     return forbidden("sitnet_reservation_in_effect");
-                } else if (exam.getState().equals(Exam.State.PUBLISHED.toString())) {
+                } else if (exam.getState() == Exam.State.PUBLISHED) {
                     // exam for reservation not started (yet?)
                     return forbidden("sitnet_reservation_in_effect");
                 }
@@ -191,13 +199,13 @@ public class EnrollController extends BaseController {
     @Restrict({@Group("ADMIN"), @Group("STUDENT")})
     public F.Promise<Result> createEnrolment(final String code, final Long id) throws MalformedURLException {
         if (!PERM_CHECK_ACTIVE) {
-            return wrapAsPromise(doCreateEnrolment(id, null));
+            return wrapAsPromise(doCreateEnrolment(id, null, ExamExecutionType.Type.PUBLIC));
         }
         final User user = getLoggedUser();
         F.Promise<Collection<String>> promise = externalAPI.getPermittedCourses(user);
         return promise.map(codes -> {
             if (codes.contains(code)) {
-                return doCreateEnrolment(id, null);
+                return doCreateEnrolment(id, null, ExamExecutionType.Type.PUBLIC);
             } else {
                 Logger.warn("Attempt to enroll for a course without permission from {}", user.toString());
                 return forbidden("sitnet_error_access_forbidden");
@@ -207,7 +215,7 @@ public class EnrollController extends BaseController {
 
     @Restrict({@Group("ADMIN"), @Group("TEACHER")})
     public Result createStudentEnrolment(Long eid, Long uid) {
-        return doCreateEnrolment(eid, uid);
+        return doCreateEnrolment(eid, uid, ExamExecutionType.Type.PRIVATE);
     }
 
     @Restrict({@Group("ADMIN"), @Group("TEACHER")})
@@ -218,8 +226,8 @@ public class EnrollController extends BaseController {
                 .eq("exam.executionType.type", ExamExecutionType.Type.PRIVATE.toString())
                 .isNull("reservation")
                 .disjunction()
-                .eq("exam.state", Exam.State.DRAFT.toString())
-                .eq("exam.state", Exam.State.SAVED.toString())
+                .eq("exam.state", Exam.State.DRAFT)
+                .eq("exam.state", Exam.State.SAVED)
                 .endJunction()
                 .disjunction()
                 .eq("exam.examOwners", user)
